requirements.txt
torch==1.8.1
torchvision==0.9.1
numpy==1.20.3
scipy==1.6.2
opencv-python==4.5.1.48
pillow==8.2.0
termcolor==1.1.0
pyyaml==5.4.1
tqdm==4.60.0
absl-py==0.12.0
gdown